% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_load.R
\name{sig_available}
\alias{sig_available}
\title{List available Signature Collection}
\usage{
sig_available()
}
\value{
a data.frame describing available datasets
}
\description{
Lists all available signature collections available in sigstash
}
\examples{

# List available datasets
sig_available()

# Load available datasets
sig_load("COSMIC_v3.3.1_SBS_GRCh38")

}
