% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sig_collection_to_sigminer}
\alias{sig_collection_to_sigminer}
\title{Sigstash Collection -> Sigminer}
\usage{
sig_collection_to_sigminer(signatures)
}
\arguments{
\item{signatures}{a signature collection}
}
\value{
a signature data.frame compatible with \code{sig} argument of sigmienr sig_fit()
}
\description{
Converts a sigstash format signature collection to a sigminer-compatible signature database.
This lets you run sigminer signature analyses sigs using sigstash
}
\details{
sigminer sig arguments allows a data.frame with rows
representing components and columns representing signatures
}
\examples{
# Load packages for signature fitting
library(sigminer)
library(quadprog)

# Load available datasets
signature_collection <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")

# Sigminer
sigminer_collection <- sig_collection_to_sigminer(signature_collection)

# Create Mock Sample Data
sample_data <- matrix(ncol = 3, runif(n = 96*3, min = 0, max = 20))

# Run Sigminer
sig_fit(sample_data, sig = sigminer_collection)


}
