% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sig_convert_channel2type}
\alias{sig_convert_channel2type}
\title{Channel to Type}
\usage{
sig_convert_channel2type(
  channel,
  sigclass = c("SBS96", "SBS1536", "ID83", "CN48", "CN80", "DBS78", "SV32", "RNA-SBS192")
)
}
\arguments{
\item{channel}{A character vector of COSMIC-style feature channels (e.g., "A[C>A]A").}

\item{sigclass}{A string indicating the class of signatures these feature sets belong to. Possible values are "SBS", "ID", "CN", "DBS", "SV", or "RNA-SBS".}
}
\value{
A character vector of the same length as \code{channel}, describing the higher-level mutation types.
}
\description{
Converts COSMIC-style channel names (e.g., "A[C>A]A") into higher-level mutation types (e.g., "C>A").
}
\examples{
doublet_channels <- c(
  "AC>CA", "AC>CG", "AC>CT", "AC>GA",
  "AC>GG", "AC>GT", "AC>TA", "AC>TG",
  "AC>TT", "AT>CA", "AT>CC", "AT>CG",
  "AT>GA", "AT>GC", "AT>TA", "CC>AA",
  "CC>AG", "CC>AT", "CC>GA", "CC>GG",
  "CC>GT", "CC>TA", "CC>TG", "CC>TT",
  "CG>AT", "CG>GC", "CG>GT", "CG>TA",
  "CG>TC", "CG>TT", "CT>AA", "CT>AC",
  "CT>AG", "CT>GA", "CT>GC", "CT>GG",
  "CT>TA", "CT>TC", "CT>TG", "GC>AA",
  "GC>AG", "GC>AT", "GC>CA", "GC>CG",
  "GC>TA", "TA>AT", "TA>CG", "TA>CT",
  "TA>GC", "TA>GG", "TA>GT", "TC>AA",
  "TC>AG", "TC>AT", "TC>CA", "TC>CG",
  "TC>CT", "TC>GA", "TC>GG", "TC>GT",
  "TG>AA", "TG>AC", "TG>AT", "TG>CA",
  "TG>CC", "TG>CT", "TG>GA", "TG>GC",
  "TG>GT", "TT>AA", "TT>AC", "TT>AG",
  "TT>CA", "TT>CC", "TT>CG", "TT>GA",
  "TT>GC", "TT>GG"
)

# Get higher-level doublet channel types
sig_convert_channel2type(doublet_channels, sigclass = "DBS78")

}
