% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sig_convert_channel_name}
\alias{sig_convert_channel_name}
\title{Convert Channel Names Between 'cosmic' and 'sigminer' Formats}
\usage{
sig_convert_channel_name(
  channel,
  from = c("cosmic", "sigminer"),
  to = c("sigminer", "cosmic")
)
}
\arguments{
\item{channel}{A character vector of channel names to be converted.}

\item{from}{A character string indicating the format to convert from. Should be either "cosmic" or "sigminer". Defaults to "cosmic".}

\item{to}{A character string indicating the format to convert to. Should be either "sigminer" or "cosmic". Defaults to "sigminer".}
}
\value{
A character vector with the converted channel names.
}
\description{
This function converts channel names between 'cosmic' and 'sigminer' formats. It modifies the capitalization of 'Kb' and replaces certain characters in the channel names as per the required format.
}
\examples{
# Convert from 'cosmic' to 'sigminer'
sig_convert_channel_name("clustered_del_1-10Kb", from = "cosmic", to = "sigminer")

# Convert from 'sigminer' to 'cosmic'
sig_convert_channel_name("clustered:del:1-10Kb", from = "sigminer", to = "cosmic")

}
