% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sig_get_valid_cosmic_types}
\alias{sig_get_valid_cosmic_types}
\title{Get Valid COSMIC Channels and Types}
\usage{
sig_get_valid_cosmic_types(sigclass)
}
\arguments{
\item{sigclass}{A string indicating the class of signatures. Run \code{\link[=sig_valid_sigclass]{sig_valid_sigclass()}} to see all options.}
}
\value{
A character vector of valid COSMIC channels or types for the specified signature class.
}
\description{
Returns a character vector of all the expected channels or types for the specified signature class (sigclass).
}
\examples{
sig_get_valid_cosmic_channels("SBS96")
sig_get_valid_cosmic_types("SBS96")
}
