% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{sig_identify_collection}
\alias{sig_identify_collection}
\title{Check the name and format of a signature collection}
\usage{
sig_identify_collection(
  signatures,
  return = c("name", "name_plus_format", "md5"),
  method = c("all", "attributes", "hash")
)
}
\arguments{
\item{signatures}{A signature dataset (in sigstash, tidy, or sigminer format) to be checked.
The signatures will be sorted before the MD5 checksum is computed.}

\item{return}{A character string specifying the type of output. Should be one of:
\itemize{
\item \code{"name"}: Returns just the dataset name.
\item \code{"name_plus_format"}: Returns both the dataset name and format in a single string.
Default is \code{"name"}.
}}

\item{method}{Method of collection identification. Should be one of:
\itemize{
\item \code{"attributes"}: identify collection based on collection_name and format attributes added by \code{\link[=sig_load]{sig_load()}}
\item \code{"hash"}: identify collection by comparing the md5 hash to a reference database of checksums.
\item \code{"all"}: identify collection using all methods available and return result of whichever is successful.
Throws an error if the methods disagree.
}}
}
\value{
If the signature collection was identified this function returns either the
dataset name  or the dataset name with the format (depending on the \code{return} argument).
If no match is found it returns \code{"Uncertain"}.
}
\description{
Identifies signature collections either by looking for attributes added by \code{\link[=sig_load]{sig_load()}}
or by comparing md5sum hash against a reference database of checksums. See the \code{method} argument documentation for details.
}
\examples{
sig_identify_collection(sig_load("COSMIC_v3.4_SBS_GRCh38"), return = "name")

sig_identify_collection(sig_load("COSMIC_v3.4_SBS_GRCh38"), return = "name_plus_format")

}
