% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_load.R
\name{sig_load}
\alias{sig_load}
\title{Load a Signature Collection}
\usage{
sig_load(dataset, format = c("sigstash", "tidy", "sigminer", "matrix"))
}
\arguments{
\item{dataset}{a valid signature collection to load. Run \code{\link[=sig_available]{sig_available()}} to list available datasets.}

\item{format}{what format should we return signature collection in (string)}
}
\value{
The specified signature collection, either as a list conforming to sigverse signature collection format, or, if \code{return_df = TRUE} as a data.frame.
}
\description{
Load a Signature Collection
}
\details{
Use the format \strong{format} argument to change return format:\tabular{ll}{
    \tab  \cr
   \strong{sigstash} \tab Signatures returned as a list of dataframes where each data.frame is a signature. \cr
   \strong{tidy} \tab Signatures returned as a tidy 4-column dataframe with signature, type, channel, fraction. \cr
   \strong{sigminer} \tab Signatures returned as a single dataframe where columns are signatures and rows are channels. Compatible with sigminer \cr
   \strong{matrix} \tab Signatures returned as a single matrix where columns are signatures and rows are channels. \cr
}
}
\examples{

# List available datasets
sig_available()

# Load available datasets
sig_load("COSMIC_v3.3.1_SBS_GRCh38")

}
