% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_load.R
\name{sig_load_annotations}
\alias{sig_load_annotations}
\title{Load Signature Annotations}
\usage{
sig_load_annotations(dataset)
}
\arguments{
\item{dataset}{a valid signature collection to load. Run \code{\link[=sig_available]{sig_available()}} to list available datasets.}
}
\value{
A data.frame containing all signature-level annotations of a given signature collection.
}
\description{
Often signature collections contain annotations describing, for example, aetiology of each signature in collections
}
\examples{

# List available datasets
sig_available()

# Load available datasets
signature_collection <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")

# Load associated annotations
signature_annotations <- sig_load_annotations("COSMIC_v3.3.1_SBS_GRCh38")

}
