% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_operations.R
\name{sig_merge}
\alias{sig_merge}
\title{Merge Signature Collections}
\usage{
sig_merge(...)
}
\arguments{
\item{...}{Signature collections to merge.}
}
\value{
A merged list of signature collections. See \href{https://github.com/selkamand/sigshared?tab=readme-ov-file#sigverse-data-types}{sigshared readme} for details.
}
\description{
Create custom signature collections using \code{\link[=sig_subset]{sig_subset()}} and \code{\link[=sig_merge]{sig_merge()}}.
Note sig_merge concatenates multiple signature collections together,
and should not be confused with \code{\link[sigstats:sig_combine]{sigstats::sig_combine()}}
which generates signature models by mathematically combining signatures from a single collection.
}
\examples{
# Load a signature collection
signature_collection <- sigstash::sig_load("COSMIC_v3.4_SBS_GRCh38")

# Subset collection to include only 2 signatures
custom_signature_collection <- sig_subset(signature_collection, c("SBS3", "SBS8"))

# Load a second signature collection
signature_collection2 <- sigstash::sig_load("COSMIC_EXPERIMENTAL_V1_SBS_HUMAN_UNFILTERED")

# Subset collection to include only 1 signature
custom_signature_collection2 <- sig_subset(signature_collection2, "temozolomide_8377cb5da514")

# Merge our 2 collections together
sig_merge(custom_signature_collection, custom_signature_collection2)
}
