% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_write.R
\name{sig_read_signatures}
\alias{sig_read_signatures}
\title{Read a Signature Collection from a File}
\usage{
sig_read_signatures(
  filepath,
  format = c("rds", "csv_tidy"),
  collection_name = NULL
)
}
\arguments{
\item{filepath}{The file path from which the signature collection should be read.}

\item{format}{The format of the file to be read. Options are:
\describe{
\item{\strong{"rds"}}{Reads the collection from an RDS file.}
\item{\strong{"csv_tidy"}}{Reads the collection from a tidy (long) CSV file and converts it to a list of signatures.}
}}

\item{collection_name}{A string specifying the signature collection name to be added as the 'collection_name' attribute
of the parsed signature object.
If NULL, name is guessed from the filename (before the first '.'). Ignored if \code{format = 'rds'}.}
}
\value{
A list of signatures in sigstash format.
}
\description{
This function reads a signature collection from a specified file, which can be either an RDS file or a tidy CSV. The collection is returned as a list of signatures in sigstash format.
}
\note{
Wide-format CSVs cannot be read back into sigverse format because they lack the necessary channel type information for accurate reconstruction (channel type is not included).
}
\examples{
\dontrun{
signatures <- sig_read_collection(filepath = "signatures.rds", format = "rds")
}
}
