% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBS96_To_SBS6.R
\name{sig_reduce_complexity_using_type}
\alias{sig_reduce_complexity_using_type}
\title{Reduce Signature Complexity}
\usage{
sig_reduce_complexity_using_type(signatures)
}
\arguments{
\item{signatures}{a sigverse signature collection}
}
\value{
a list conforming to sigverse signature collection format
}
\description{
Reduce complex signatures into simpler forms by converting 'channel' level signatures into 'type' level signatures Takes complex signatures like SBS96 and turns them into SBS6 signatures.
}
