% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sig_valid_sigclass}
\alias{sig_valid_sigclass}
\title{List Valid Signature Classes}
\usage{
sig_valid_sigclass()
}
\value{
vector of sigverse supported signature catalogue feature-spaces.
}
\description{
List sigverse supported signature feature-spaces.
}
\examples{
sig_valid_sigclass()
}
