% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_write.R
\name{sig_write_signatures}
\alias{sig_write_signatures}
\title{Write a Signature Collection to a File}
\usage{
sig_write_signatures(
  signatures,
  filepath,
  format = c("rds", "csv_tidy", "csv_wide")
)
}
\arguments{
\item{signatures}{A signature collection in sigstash format. This should be a list of signatures.}

\item{filepath}{The file path where the signature collection should be saved.}

\item{format}{The format in which the signatures should be saved. Options are:
\describe{
\item{\strong{"rds"}}{Serializes the list in RDS format.}
\item{\strong{"csv_tidy"}}{Converts the collection to a tidy (long) data frame and stores it as a CSV file.}
\item{\strong{"csv_wide"}}{Converts the collection to a wide format (with channels as row names) and stores it as a CSV file.}
}}
}
\value{
Invisibly returns \code{NULL}. This function is called for its side effects of writing data to a file.
}
\description{
This function writes a signature collection, in sigstash format, to a specified file. The collection can be saved either as an RDS object, a tidy CSV, or a wide CSV.
}
\examples{
\dontrun{
signature_collection <- sigstash::sig_load("COSMIC_v3.4_SBS_GRCh38")
sig_write_collection(signature_collection, filepath = "signatures.rds", format = "rds")
}
}
