% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{sig_get_valid_cosmic_types}
\alias{sig_get_valid_cosmic_types}
\title{Get Valid COSMIC Channels and Types}
\usage{
sig_get_valid_cosmic_types(
  sigclass = c("SBS96", "SBS1536", "ID83", "CN48", "DBS78", "SV32", "RNA-SBS192")
)
}
\arguments{
\item{sigclass}{A string indicating the class of signatures. Possible values are "SBS96", "SBS1536", "ID83", "CN48", "DBS78", "SV32", or "RNA-SBS192".}
}
\value{
A character vector of valid COSMIC channels or types for the specified signature class.
}
\description{
Returns a character vector of all the expected channels or types for the specified signature class (sigclass).
}
\examples{
sig_get_valid_cosmic_channels("SBS96")
sig_get_valid_cosmic_types("SBS96")
}
