% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{sig_identify_collection}
\alias{sig_identify_collection}
\title{Check if a signature collection matches a known dataset}
\usage{
sig_identify_collection(
  signatures,
  return = c("name", "name_plus_format", "md5")
)
}
\arguments{
\item{signatures}{A signature dataset (in sigstash, tidy, or sigminer format) to be checked.
The signatures will be sorted before the MD5 checksum is computed.}

\item{return}{A character string specifying the type of output. Should be one of:
\itemize{
\item \code{"name"}: Returns just the dataset name.
\item \code{"name_plus_format"}: Returns both the dataset name and format in a single string.
Default is \code{"name"}.
}}
}
\value{
If the signature collection matches a known dataset, the function returns either
the dataset name or the dataset name with the format. If no match is found,
it returns \code{"Uncertain"}.
}
\description{
The \code{sig_identify_collection} function checks whether a given signature collection
matches any precomputed dataset by computing its MD5 checksum and comparing it
with a reference set of checksums. The reference checksums are loaded from
precomputed data. The function can return either the dataset name or the
dataset name along with the format (e.g., 'tidy', 'sigstash', 'sigminer').
}
\examples{
# Assuming 'my_signatures' is a signature dataset:
sig_identify_collection(sig_load("COSMIC_v3.4_SBS_GRCh38"), return = "name")

# Example returning the dataset name along with its format:
sig_identify_collection(sig_load("COSMIC_v3.4_SBS_GRCh38"), return = "name_plus_format")

}
